Allow auto-creating agents in bits ask.
