+++ src/commands/bits.rs
@@ -1,8 +1,7 @@
 use anyhow::Result;
 
+use crate::commands::lassie;
 use crate::config::Config;
-
-const LASSIE_BASE: &str = "/api/unstable/lassie-ng/v1";
 
 /// Ask Datadog Bits AI a natural-language question.
 ///
@@ -29,7 +28,7 @@ pub async fn ask(
     let agent_id = match agent_id {
         Some(id) if !id.is_empty() => id,
         _ => {
-            resolve_agent_id(
+            lassie::resolve_agent_id(
                 &app_base,
                 cfg.access_token.as_deref(),
                 cfg.api_key.as_deref(),
@@ -95,7 +94,10 @@ async fn send_turn(
     session_id: Option<String>,
     stream: bool,
 ) -> Result<Option<String>> {
-    let url = format!("{app_base}{LASSIE_BASE}/agents/{agent_id}/messages");
+    let url = format!(
+        "{app_base}{}/agents/{agent_id}/messages",
+        lassie::LASSIE_BASE
+    );
 
     let mut body = serde_json::json!({ "input": query, "stream": stream });
     if let Some(ref sid) = session_id {
@@ -247,151 +249,15 @@ fn extract_text(val: &serde_json::Value) -> String {
     parts.join("\n")
 }
 
-/// Resolve the first available Bits AI agent ID from the API.
-///
-/// If `auto_create` is true and no agents are found, creates a new agent.
-#[cfg(not(target_arch = "wasm32"))]
-async fn resolve_agent_id(
-    app_base: &str,
-    access_token: Option<&str>,
-    api_key: Option<&str>,
-    app_key: Option<&str>,
-    auto_create: bool,
-) -> Result<String> {
-    let url = format!("{app_base}{LASSIE_BASE}/agents?limit=1");
-    let client = reqwest::Client::new();
-    let req = client.get(&url).header("Accept", "application/json");
-
-    let req = req.header("User-Agent", crate::useragent::get());
-    let req = if let Some(token) = access_token {
-        req.header("Authorization", format!("Bearer {token}"))
-    } else if let (Some(ak), Some(apk)) = (api_key, app_key) {
-        req.header("DD-API-KEY", ak)
-            .header("DD-APPLICATION-KEY", apk)
-    } else {
-        anyhow::bail!("no authentication configured");
-    };
-
-    let resp = req
-        .send()
-        .await
-        .map_err(|e| anyhow::anyhow!("Failed to list Bits AI agents: {e}"))?;
-
-    if !resp.status().is_success() {
-        let status = resp.status();
-        let body = resp.text().await.unwrap_or_default();
-        anyhow::bail!("GET /agents failed (HTTP {status}): {body}");
-    }
-
-    let val: serde_json::Value = resp
-        .json()
-        .await
-        .map_err(|e| anyhow::anyhow!("Failed to parse agents response: {e}"))?;
-
-    let agents = val.as_array().ok_or_else(|| {
-        anyhow::anyhow!(
-            "Unexpected response format from Bits AI agents API.\n\
-                 Expected a JSON array but got: {}\n\
-                 This may indicate an API version mismatch — please report this at\n\
-                 https://github.com/DataDog/pup/issues if the issue persists.",
-            serde_json::to_string(&val).unwrap_or_else(|_| "<unparseable>".to_string())
-        )
-    })?;
-
-    if agents.is_empty() {
-        if auto_create {
-            eprintln!("No Bits AI agents found — creating one automatically...");
-            return create_agent(app_base, access_token, api_key, app_key).await;
-        }
-        anyhow::bail!(
-            "No Bits AI agents found in your Datadog organization.\n\
-             \n\
-             To fix this, create a Bits AI agent at:\n\
-               https://app.datadoghq.com/actions/agents/create\n\
-             \n\
-             Once created, `pup bits ask` will auto-discover it.\n\
-             Alternatively, pass --auto-create to have pup create one for you,\n\
-             or pass an existing agent ID with --agent-id."
-        );
-    }
-
-    let id = agents[0]
-        .get("id")
-        .and_then(|v| v.as_str())
-        .ok_or_else(|| anyhow::anyhow!("Agent missing 'id' field"))?;
-
-    Ok(id.to_string())
-}
-
-/// Create a new Bits AI agent and return its ID.
-#[cfg(not(target_arch = "wasm32"))]
-async fn create_agent(
-    app_base: &str,
-    access_token: Option<&str>,
-    api_key: Option<&str>,
-    app_key: Option<&str>,
-) -> Result<String> {
-    let url = format!("{app_base}{LASSIE_BASE}/agents");
-    let body = serde_json::json!({
-        "name": "Pup CLI Agent",
-        "description": "Auto-created by pup CLI for bits ask",
-    });
-
-    let client = reqwest::Client::new();
-    let req = client
-        .post(&url)
-        .header("Content-Type", "application/json")
-        .header("Accept", "application/json")
-        .header("User-Agent", crate::useragent::get());
-
-    let req = if let Some(token) = access_token {
-        req.header("Authorization", format!("Bearer {token}"))
-    } else if let (Some(ak), Some(apk)) = (api_key, app_key) {
-        req.header("DD-API-KEY", ak)
-            .header("DD-APPLICATION-KEY", apk)
-    } else {
-        anyhow::bail!("no authentication configured");
-    };
-
-    let resp = req
-        .json(&body)
-        .send()
-        .await
-        .map_err(|e| anyhow::anyhow!("Failed to create Bits AI agent: {e}"))?;
-
-    if !resp.status().is_success() {
-        let status = resp.status();
-        let err_body = resp.text().await.unwrap_or_default();
-        anyhow::bail!("Failed to create Bits AI agent (HTTP {status}): {err_body}");
-    }
-
-    let val: serde_json::Value = resp
-        .json()
-        .await
-        .map_err(|e| anyhow::anyhow!("Failed to parse create-agent response: {e}"))?;
-
-    let id = val
-        .get("id")
-        .and_then(|v| v.as_str())
-        .ok_or_else(|| anyhow::anyhow!("Create-agent response missing 'id' field"))?;
-
-    eprintln!("Created Bits AI agent: {id}");
-    Ok(id.to_string())
-}
-
-/// Attach auth headers to a request builder.
+/// Convenience wrapper that extracts credentials from `Config`.
 #[cfg(not(target_arch = "wasm32"))]
 fn add_auth(req: reqwest::RequestBuilder, cfg: &Config) -> Result<reqwest::RequestBuilder> {
-    let req = req.header("User-Agent", crate::useragent::get());
-    if let Some(token) = &cfg.access_token {
-        return Ok(req.header("Authorization", format!("Bearer {token}")));
-    }
-    if let (Some(ak), Some(apk)) = (&cfg.api_key, &cfg.app_key) {
-        return Ok(req
-            .header("DD-API-KEY", ak.as_str())
-            .header("DD-APPLICATION-KEY", apk.as_str()));
-    }
-    anyhow::bail!("no authentication configured")
+    lassie::add_auth(
+        req,
+        cfg.access_token.as_deref(),
+        cfg.api_key.as_deref(),
+        cfg.app_key.as_deref(),
+    )
 }
 
 // ---------------------------------------------------------------------------
